Move shopify initialization to root element and update styling
