+++ package.json
@@ -31,6 +31,7 @@
     "react-apollo": "^3.1.5",
     "react-burger-menu": "^2.6.10",
     "react-dom": "^16.8.6",
+    "react-ga": "^3.2.0",
     "react-helmet": "^6.0.0",
     "react-router-dom": "^5.0.1",
     "react-scripts": "^3.4.3",
