+++ src/templates/product.js
@@ -17,13 +17,6 @@ class ProductTemplate extends React.Component {
    * @return {object} product template object
    */
   render() {
-    const fetch = require('isomorphic-fetch');
-    const shopify = require('shopify-buy');
-
-    const client = shopify.buildClient({
-      domain: 'wynne-the-pooh.myshopify.com',
-      storefrontAccessToken: '1e5568ef2b02e41e3e1a1755f1af254a',
-    }, fetch);
 
     const product = this.props.data.markdownRemark;
     const siteTitle = get(this.props, 'data.site.siteMetadata.title');
@@ -36,7 +29,7 @@ class ProductTemplate extends React.Component {
         }));
 
     return (
-      <Layout location={this.props.location} title={siteTitle} client={client}>
+      <Layout location={this.props.location} title={siteTitle}>
         <ProductPage>
 
           <Link
